+++ src/ragger/navigator/navigator.py
@@ -74,7 +74,6 @@ class NavInsID(Enum):
     USE_CASE_SUB_SETTINGS_NEXT = auto()
     USE_CASE_CHOICE_CONFIRM = auto()
     USE_CASE_CHOICE_REJECT = auto()
-    USE_CASE_STATUS_WAIT = auto()
     USE_CASE_STATUS_DISMISS = auto()
     USE_CASE_REVIEW_TAP = auto()
     USE_CASE_REVIEW_PREVIOUS = auto()